working on TeacherController and bug of Bignumbers

+++ prisma/tables.sql
@@ -1,6 +1,6 @@
 -- Teachers Table
 CREATE TABLE Teachers (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     email VARCHAR(255) UNIQUE NOT NULL,
     phone VARCHAR(255) UNIQUE NOT NULL,
@@ -9,7 +9,7 @@ CREATE TABLE Teachers (
 
 -- Centers Table
 CREATE TABLE Centers (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     location VARCHAR(255) NOT NULL,
     phone VARCHAR(255) UNIQUE NOT NULL
@@ -17,16 +17,16 @@ CREATE TABLE Centers (
 
 -- Teacher-Centers (Many-to-Many Relationship)
 CREATE TABLE Teacher_Centers (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    teacher_id BIGINT NOT NULL,
-    center_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    teacher_id INT NOT NULL,
+    center_id INT NOT NULL,
     FOREIGN KEY (teacher_id) REFERENCES Teachers(id) ON DELETE CASCADE,
     FOREIGN KEY (center_id) REFERENCES Centers(id) ON DELETE CASCADE
 );
 
 -- Parents Table
 CREATE TABLE Parents (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     phone VARCHAR(255) UNIQUE NOT NULL,
     email VARCHAR(255) UNIQUE,
@@ -35,20 +35,20 @@ CREATE TABLE Parents (
 
 -- Students Table
 CREATE TABLE Students (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     email VARCHAR(255) UNIQUE,
     section ENUM('first_sec', 'second_sec_scientific', 'second_sec_literary', 'third_sec') NOT NULL,
-    center_id BIGINT,
+    center_id INT,
     phone VARCHAR(255) UNIQUE NOT NULL,
     password VARCHAR(255) NOT NULL,
     FOREIGN KEY (center_id) REFERENCES Centers(id) ON DELETE SET NULL
 );
 
 CREATE TABLE Student_Parents (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    student_id BIGINT NOT NULL,
-    parent_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    student_id INT NOT NULL,
+    parent_id INT NOT NULL,
     relationship ENUM('father', 'mother', 'guardian') NOT NULL, -- Defines role
     FOREIGN KEY (student_id) REFERENCES Students(id) ON DELETE CASCADE,
     FOREIGN KEY (parent_id) REFERENCES Parents(id) ON DELETE CASCADE,
@@ -56,11 +56,11 @@ CREATE TABLE Student_Parents (
 );
 -- Sessions Table
 CREATE TABLE Sessions (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    teacher_id BIGINT,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    teacher_id INT,
     date DATETIME NOT NULL,
     topic VARCHAR(255) NOT NULL,
-    center_id BIGINT,
+    center_id INT,
     section ENUM('first_sec', 'second_sec_scientific', 'second_sec_literary', 'third_sec') NOT NULL,
     FOREIGN KEY (teacher_id) REFERENCES Teachers(id) ON DELETE SET NULL,
     FOREIGN KEY (center_id) REFERENCES Centers(id) ON DELETE SET NULL
@@ -68,8 +68,8 @@ CREATE TABLE Sessions (
 
 -- Homeworks Table
 CREATE TABLE Homeworks (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    session_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    session_id INT NOT NULL,
     description VARCHAR(255) NOT NULL,
     due_date DATETIME NOT NULL,
     FOREIGN KEY (session_id) REFERENCES Sessions(id) ON DELETE cascade
@@ -77,8 +77,8 @@ CREATE TABLE Homeworks (
 
 -- Quizzes Table
 CREATE TABLE Quizzes (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    session_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    session_id INT NOT NULL,
     max_score INTEGER NOT NULL,
     date DATETIME NOT NULL,
     FOREIGN KEY (session_id) REFERENCES Sessions(id) ON DELETE cascade
@@ -86,22 +86,22 @@ CREATE TABLE Quizzes (
 
 -- Trials Table
 CREATE TABLE Trials (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id INT AUTO_INCREMENT PRIMARY KEY,
     description VARCHAR(255) NOT NULL,
     date DATETIME NOT NULL,
     max_score INTEGER NOT NULL,
     section ENUM('first_sec', 'second_sec_scientific', 'second_sec_literary', 'third_sec') NOT NULL,
-    teacher_id BIGINT,
-    center_id BIGINT,
+    teacher_id INT,
+    center_id INT,
     FOREIGN KEY (teacher_id) REFERENCES Teachers(id) ON DELETE SET NULL,
     foreign key (center_id) references Centers(id) on delete set null
 );
 
 -- Student-Sessions Table (Many-to-Many)
 CREATE TABLE Student_Sessions (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    session_id BIGINT NOT NULL,
-    student_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    session_id INT NOT NULL,
+    student_id INT NOT NULL,
     notes VARCHAR(255),
     attendance_status ENUM('present', 'absent') DEFAULT 'absent',
     FOREIGN KEY (session_id) REFERENCES Sessions(id) ON DELETE CASCADE,
@@ -110,9 +110,9 @@ CREATE TABLE Student_Sessions (
 
 -- Student-Trials Table (Many-to-Many)
 CREATE TABLE Student_Trials (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    trial_id BIGINT NOT NULL,
-    student_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    trial_id INT NOT NULL,
+    student_id INT NOT NULL,
     result VARCHAR(255) NOT NULL,
     FOREIGN KEY (trial_id) REFERENCES Trials(id) ON DELETE CASCADE,
     FOREIGN KEY (student_id) REFERENCES Students(id) ON DELETE CASCADE
@@ -120,9 +120,9 @@ CREATE TABLE Student_Trials (
 
 -- Student-Homeworks Table (Many-to-Many)
 CREATE TABLE Student_Homeworks (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    homework_id BIGINT NOT NULL,
-    student_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    homework_id INT NOT NULL,
+    student_id INT NOT NULL,
     status VARCHAR(255) NOT NULL,
     submission_date DATETIME NOT NULL ,
     FOREIGN KEY (homework_id) REFERENCES Homeworks(id) ON DELETE CASCADE,
@@ -132,9 +132,9 @@ CREATE TABLE Student_Homeworks (
 
 -- Student-Quizzes Table (Many-to-Many)
 CREATE TABLE Student_Quizzes (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    quiz_id BIGINT NOT NULL,
-    student_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    quiz_id INT NOT NULL,
+    student_id INT NOT NULL,
     score INTEGER NOT NULL,
     FOREIGN KEY (quiz_id) REFERENCES Quizzes(id) ON DELETE CASCADE,
     FOREIGN KEY (student_id) REFERENCES Students(id) ON DELETE CASCADE
